Changed Customer id from long to int. No need for long as int is big enough for any of the given problems.

+++ src/com/vr/coursework/helpers/Customer.java
@@ -5,9 +5,9 @@ import java.awt.geom.Point2D;
 public class Customer extends Point2D.Double{
 
 	// Requirements of the customer (number to be delivered)
-	private static long counter = 0;
+	private static int counter = 0;
 	public int c;
-	public long id;
+	public int id;
 	
 	public Customer(int x, int y, int requirement){
 		this.id = counter;
